This PR reduces allocations in `PageParameters.mergeWith`. As a side-effect, it increases throughput 4x.

It avoids allocating new collections where possible and attempts to initialize `namedParameters` with the correct size.

This method is called more than 1000 times per request in my application because we have a lot of mounted pages and resources and it is called for every single one of them.

**Master**
Benchmark    |                      Mode |  Cnt    |      Score     |      Units
------------ | ------------- |  ------------- |  --: | -------------
ParameterBenchmarks.mergeWith                               |    thrpt  |  3 |  **1746723,436**  | ops/s
ParameterBenchmarks.mergeWith:·gc.alloc.rate.norm           |    thrpt  |  3 |     1600,190  | B/op
ParameterBenchmarks.mergeWith:·gc.churn.G1_Eden_Space.norm  |    thrpt  |  3 |     1602,183  | B/op

**Branch**
Benchmark    |                      Mode |  Cnt    |      Score     |      Units
------------ | ------------- |  ------------- |  --: | -------------
ParameterBenchmarks.mergeWith                                |   thrpt |   3 | **7223753,855** |  ops/s
ParameterBenchmarks.mergeWith:·gc.alloc.rate.norm            |   thrpt |   3 |     320,046 |  B/op
ParameterBenchmarks.mergeWith:·gc.churn.G1_Eden_Space.norm   |   thrpt |   3 |     320,760 |  B/op

https://issues.apache.org/jira/browse/WICKET-6833